+++ src/components/AudioVolumeControl.tsx
@@ -39,6 +39,7 @@ export const useComponentStyles = makeStyles((theme: any) => {
     volumeControlContainer: {
       position: 'absolute',
       display: 'none',
+      zIndex: 10,
       [theme.breakpoints.up('sm')]: {
         display: 'flex',
         height: '60px',
@@ -55,6 +56,7 @@ interface IPlayerVolume {
 interface IAudioDownloadsControl {
   playerColors: IAudioPlayerColors;
   volume: IPlayerVolume;
+  muted?: Boolean | null;
   muteAudio: () => void;
   unmuteAudio: () => void;
   changeAudioVolume: (value: any) => void;
@@ -63,6 +65,7 @@ interface IAudioDownloadsControl {
 }
 
 export const AudioVolumeControl: React.FunctionComponent<IAudioDownloadsControl> = ({
+  muted = null,
   muteAudio,
   unmuteAudio,
   classNames = {},
@@ -80,6 +83,7 @@ export const AudioVolumeControl: React.FunctionComponent<IAudioDownloadsControl>
   const toggleVolumeSlider = (value: boolean) => () => {
     openVolumeSlider(value);
   };
+
   return (
     <Grid
       item={true}
@@ -91,17 +95,22 @@ export const AudioVolumeControl: React.FunctionComponent<IAudioDownloadsControl>
         <VolumeUpIcon
           fontSize="large"
           className={cx(classes.icon, classNames.volumeIcon)}
-          onClick={muteAudio}
+          onClick={typeof muted !== 'boolean' ? muteAudio : () => {}}
         />
       ) : (
         <VolumeOffIcon
           fontSize="large"
           className={cx(classes.icon, classNames.volumeIcon)}
-          onClick={unmuteAudio}
+          onClick={typeof muted !== 'boolean' ? unmuteAudio : () => {}}
         />
       )}
       {volumeSlider && (
-        <Paper className={cx(classes.volumeControlContainer)}>
+        <Paper
+          className={cx(
+            classes.volumeControlContainer,
+            classNames.volumeSliderContainer
+          )}
+        >
           <Slider
             orientation="vertical"
             aria-labelledby="volume-control"