+++ stories/index.stories.tsx
@@ -76,6 +76,14 @@ storiesOf('Material Ui', module)
       const download = boolean('download', true);
       const debug = boolean('debug', true);
       const loop = boolean('loop', true);
+      const muted = select(
+        'muted',
+        { null: null, true: true, false: false },
+        null
+      );
+      const volume = boolean('volume', true);
+      const displaySlider = boolean('displaySlider', true);
+      const displayCloseButton = boolean('displayCloseButton', false);
       const onFinished = (event): void => {
         eventTriggered('onFinish');
       };
@@ -93,9 +101,11 @@ storiesOf('Material Ui', module)
             width={width}
             variation={variation}
             download={download}
+            volume={volume}
             loop={loop}
             order={order}
             spacing={spacing}
+            muted={muted}
             debug={debug}
             src={src}
             onFinished={onFinished}
@@ -103,6 +113,8 @@ storiesOf('Material Ui', module)
             onPlayed={onPlayed}
             time={time}
             timePosition={timePosition}
+            displaySlider={displaySlider}
+            displayCloseButton={displayCloseButton}
           />
         </ThemeProvider>
       );
@@ -273,6 +285,26 @@ const useStyles = makeStyles(
     return (
       <ThemeProvider theme={muiTheme}>
         <AudioPlayer src="https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3" />
+      </ThemeProvider>
+    );
+  })
+  .add('Small AudioPlayer', () => {
+    const displayCloseButton = boolean('displayCloseButton', true);
+    const width = text('width', '58px');
+    // tslint:disable-next-line
+    const onClose = () => console.log('closed');
+
+    return (
+      <ThemeProvider theme={muiTheme}>
+        <AudioPlayer
+          src="https://www.soundhelix.com/examples/mp3/SoundHelix-Song-1.mp3"
+          width={width}
+          variation="primary"
+          displaySlider={false}
+          volume={false}
+          displayCloseButton={displayCloseButton}
+          onClose={onClose}
+        />
       </ThemeProvider>
     );
   })